pyproject.toml
[tool.poetry]
name = "vg-mlops23"
version = "0.1.0"
description = ""
authors = ["Vigolod <109913264+Vigolod@users.noreply.github.com>"]
license = "GNU-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.8.2"
numpy = "^1.26.2"
hydra-core = "^1.3.2"
tqdm = "^4.66.1"
torch = "^2.1.2"
torchvision = "^0.16.2"
safetensors = "^0.4.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
dvc = "^3.31.2"
dvc-gdrive = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["vg-mlops23"]
profile = "black"
line_length = 90
lines_after_imports = 2
